+++ src/test/java/lotto/domain/strategy/LottoNumberAutoGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.stragegy;
+package lotto.domain.strategy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
